+++ python/wintask/__init__.py
@@ -1,3 +1,4 @@
+from .errors import TaskNameError
 from .scheduler import TaskScheduler
 from .triggers import DailyTrigger, Month, MonthlyTrigger, Weekday, WeeklyTrigger
 
@@ -5,6 +6,7 @@ __all__ = [
 	"DailyTrigger",
 	"Month",
 	"MonthlyTrigger",
+	"TaskNameError",
 	"TaskScheduler",
 	"Weekday",
 	"WeeklyTrigger",